Moved inference out of types.
