+++ mathesar/utils/tables.py
@@ -2,7 +2,7 @@ from sqlalchemy import MetaData
 
 from db.tables.operations.create import create_mathesar_table
 from db.tables.operations.select import get_oid_from_table
-from db.types.inference import infer_table_column_types
+from db.tables.operations.infer_types import infer_table_column_types
 from mathesar.database.base import create_mathesar_engine
 from mathesar.imports.csv import create_table_from_csv
 from mathesar.models import Table
